+++ sdk/core/azure-core-http-test-common/src/main/java/com/azure/core/http/test/common/models/package-info.java
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+/**
+ * Package containing models that are used by HTTP client implementations to test against.
+ */
+package com.azure.core.http.test.common.models;