+++ sdk/core/azure-core-test/src/main/java/module-info.java
@@ -3,6 +3,10 @@
 
 module com.azure.core.test {
     requires transitive com.azure.core;
+    requires transitive com.azure.core.http.jdk.httpclient;
+    requires transitive com.azure.http.netty;
+    requires transitive com.azure.core.http.okhttp;
+    requires transitive com.azure.core.http.vertx;
 
     requires org.junit.jupiter.api;
     requires org.junit.jupiter.params;
@@ -10,15 +14,10 @@ module com.azure.core.test {
     requires reactor.test;
     requires java.management;
     requires java.net.http;
-    requires javax.servlet.api;
     requires reactor.netty.http;
     requires reactor.netty.core;
     requires io.netty.codec.http;
     requires org.apache.commons.compress;
-    requires org.eclipse.jetty.server;
-    requires org.eclipse.jetty.servlet;
-    requires org.eclipse.jetty.util;
-    requires org.eclipse.jetty.security;
 
     exports com.azure.core.test;
     exports com.azure.core.test.annotation;
@@ -31,7 +30,6 @@ module com.azure.core.test {
     opens com.azure.core.test.models to com.fasterxml.jackson.databind;
     opens com.azure.core.test.implementation
         to com.fasterxml.jackson.databind, com.azure.core, org.junit.platform.commons;
-    opens com.azure.core.test.implementation.entities to com.fasterxml.jackson.databind, com.azure.core;
     opens com.azure.core.test to com.azure.core, com.fasterxml.jackson.databind, org.junit.platform.commons;
     opens com.azure.core.test.junitextensions
         to com.azure.core, com.fasterxml.jackson.databind, org.junit.platform.commons;